+++ axicor-lab/src/layout/mod.rs
@@ -1,3 +1,4 @@
-pub mod tree;
+pub mod data;
+pub mod behavior;
 pub mod systems;
-pub mod interaction;
+pub mod input;